backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv

# GitHub repo handling
gitpython

# RAG + embeddings
sentence-transformers
langchain
langchain-text-splitters
langchain-groq

# Vector database
chromadb
pinecone

# Utilities
pydantic
